Added grammar change after operator feature
Allows things like having the | in a Spin method header that leads the local
variables and a | in an array size specifer for a local variable on the
same line to have different precedences.

Also changed blockdef object to have a pointer to a distinct grammardef
object.

Also added half-implemented symbol scope linked list, since it was too
entangled/I was too lazy to commit it separately